Realizada a adaptação do código com blue e isort, todas as funções do bot.py estão em core.py
