--- .github/ISSUE_TEMPLATE/feature-request.md
@@ -1,20 +1,0 @@
----
-name: Feature Request
-about: Suggest an idea for this project
-title: Add a function to create a cat button
-labels: ''
-assignees: ''
-
----
-
-**Is your feature request related to a problem? Please describe.**
-A clear and concise description of what the problem is. Ex. I'm always frustrated when [...]
-
-**Describe the solution you'd like**
-A clear and concise description of what you want to happen.
-
-**Describe alternatives you've considered**
-A clear and concise description of any alternative solutions or features you've considered.
-
-**Additional context**
-Add any other context or screenshots about the feature request here.
