--- .vscode/settings.json
@@ -1,5 +1,0 @@
-{
-    "makefile.extensionOutputFolder": "./.vscode",
-    "python.linting.pycodestyleEnabled": true,
-    "python.linting.enabled": true
-}
