+++ app/tax/brazil/impostos.py
@@ -1,8 +1,10 @@
 from telegram import Update
-from telegram.ext import ConversationHandler, ContextTypes
+from telegram.ext import ContextTypes, ConversationHandler
 
 
-async def calculo_imposto_de_renda(update: Update, context: ContextTypes.DEFAULT_TYPE):
+async def calculo_imposto_de_renda(
+    update: Update, context: ContextTypes.DEFAULT_TYPE
+):
     """
     Essa função está recebendo a mensagem do usuário a qual foi passada na
     função 'renda'. Futuramente, nessa função pode ser retornado um cálculo
@@ -13,8 +15,10 @@ async def calculo_imposto_de_renda(update: Update, context: ContextTypes.DEFAULT
         constante que serve para encerrar uma conversa.
     """
     user = update.message.text
-    await update.message.reply_text(f"""
+    await update.message.reply_text(
+        f"""
     Entendi, então sua renda é {user}
-    """)
+    """
+    )
 
     return ConversationHandler.END
