+++ pyproject.toml
@@ -14,7 +14,8 @@ python-dotenv = "^0.21.0"
 [tool.poetry.dev-dependencies]
 
 [tool.poetry.group.dev.dependencies]
-pycodestyle = "^2.9.1"
+blue = "^0.9.1"
+isort = "^5.10.1"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
